# Task-2759 optimize AccessControl function for general isContentAvailable question