--- a/app/Models/Bible/Bible.php
+++ b/app/Models/Bible/Bible.php
@@ -433,20 +433,13 @@ class Bible extends Model
 
     public function scopeIsContentAvailable(Builder $query, Collection $access_group_ids)
     {
-        return $query->whereExists(function ($query) use ($access_group_ids) {
-            return $query->select(\DB::raw(1))
-                ->from('sys_license_group_access_groups_view as lgag')
-                ->join(
-                    'bible_filesets as abf',
-                    function ($join) {
-                        $join->on('abf.license_group_id', '=', 'lgag.lg_id')
-                            ->where('abf.content_loaded', true)
-                            ->where('abf.archived', false);
-                    }
-                )->join('bible_fileset_connections as bfc', 'abf.hash_id', 'bfc.hash_id')
-                ->whereColumn('bibles.id', '=', 'bfc.bible_id')
-                ->whereIn('lgag.access_group_id', $access_group_ids);
-        });
+        $sub = BibleFileset::query()
+            ->selectRaw('1')
+            ->isContentAvailable($access_group_ids)
+            ->join('bible_fileset_connections as bfc', 'bible_filesets.hash_id', '=', 'bfc.hash_id')
+            ->whereColumn('bfc.bible_id', 'bibles.id');
+
+        return $query->whereExists($sub);
     }
     public function scopeIsTimingInformationAvailable($query)
     {