--- a/app/Models/Bible/BibleFilesetConnection.php
+++ b/app/Models/Bible/BibleFilesetConnection.php
@@ -77,20 +77,11 @@ class BibleFilesetConnection extends Model
 
     public function scopeIsContentAvailable(Builder $query, Collection $access_group_ids)
     {
-        return $query->whereExists(function ($query) use ($access_group_ids) {
-            return $query->select(\DB::raw(1))
-                ->from('sys_license_group_access_groups_view as lgag')
-                ->join(
-                    'bible_filesets as abf',
-                    function ($join) {
-                        $join->on('abf.license_group_id', '=', 'lgag.lg_id')
-                            ->where('abf.content_loaded', true)
-                            ->where('abf.archived', false)
-                            ;
-                    }
-                )
-                ->whereColumn('abf.hash_id', '=', 'bible_fileset_connections.hash_id')
-                ->whereIn('lgag.access_group_id', $access_group_ids);
-        });
+        $sub = BibleFileset::query()
+            ->selectRaw('1')
+            ->isContentAvailable($access_group_ids)
+            ->whereColumn('bible_filesets.hash_id', '=', 'bible_fileset_connections.hash_id');
+
+        return $query->whereExists($sub);
     }
 }