# Description
Added an optimization that checks whether content is available for a given API user key. If a fileset’s permission pattern exists and is not CUSTOM, the API now validates access through the permission-pattern/access-group relationship. If the permission pattern is NULL or CUSTOM, it falls back to the existing check against access_group_filesets.
This change speeds up the general IsContentAvailable query.

For Access Control on a specific fileset, the API still uses the view SysLicenseGroupAccessGroups so that Blimp and the API rely on the same single source of truth.

## Issue Link
[User Story 2759](https://dev.azure.com/keyholesoftware/FCBH%20-%20Bible%20Brain/_workitems/edit/2759): Public API: optimize AccessControl function for general isContentAvailable question

## How Do I QA This
Run all [M] Postman tests and they should all pass.

## Notes on Performance

Below are three examples that show the improvement, all using the bible.is API key. Before and after screenshots are attached.


Endpoint | Purpose | Before | After
-- | -- | -- | --
api/bibles/search/v?v=4&key=API_USER_KEY | General isContentAvailable query (v4_bible.search) | ~2 s | < 100 ms
api/languages/search/arabic?v=4&key=API_USER_KEY | Language search with translations | ~4 s  | < 100 ms


1. `Bibles search (v4_bible.search)` - It is an endpoint related to the general question of whether content is available
api/bibles/search/v?v=4&key=API_USER_KEY

After
![Screenshot from 2025-05-16 18-07-25](https://github.com/user-attachments/assets/8bd8fe70-0aae-4bde-a0c7-3a7ce264276d)
Before
![Screenshot from 2025-05-16 18-07-04](https://github.com/user-attachments/assets/0f31f33e-66d7-427c-aab8-9ec25c2e8d62)

2. `Language search languages with translations` - It is an endpoint related to the general question of whether content is available
api/languages/search/arabic?v=4&key=API_USER_KEY

Before
![Screenshot from 2025-05-16 19-21-05](https://github.com/user-attachments/assets/01df3033-ea65-4a34-8f7d-c7e033a4c58b)
After
![Screenshot from 2025-05-16 19-20-40](https://github.com/user-attachments/assets/87b67ba7-1c81-4e86-9adb-99c054a4768a)